fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9480 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/gui/tracker/TrackerValueDesc.cpp
@@ -153,7 +153,7 @@ TrackerValueDesc::setAggregationSpan(SUMOTime as) {
         while (i!=myValues.end()) {
             myTmpLastAggValue = 0;
             myValidNo = 0;
-            for (size_t j=0; j<myAggregationInterval&&i!=myValues.end(); j++, ++i) {
+            for (int j=0; j<myAggregationInterval&&i!=myValues.end(); j++, ++i) {
                 if ((*i)!=myInvalidValue) {
                     myTmpLastAggValue += (*i);
                     myValidNo++;
